fixup docs
